スタンプ関連を抽出してファイルを分割
